choral_rss_bot/pyproject.toml
[project]
name = "choral-rss-bot"
version = "0.1.0"
description = "A Discord bot to fetch and post Choral music RSS feeds with LLM summarization."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.11",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
    "google-genai>=0.2.0",
    "rich>=13.9.4",
    "python-dateutil>=2.8.2",
    "google-cloud-firestore>=2.14.0",
    "google-adk>=0.2.0",
    "playwright>=1.49.0",
    "google-cloud-aiplatform[agent_engines,adk]>=1.112",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.0.0",
]
